src/python-processor/requirements.txt
boto3
psycopg2-binary
opentelemetry-api
Flask
prometheus-flask-exporter
